+++ src/main/java/com/inventario/inventario/service/ProductoService.java
@@ -20,7 +20,7 @@ public class ProductoService {
         return productoRepository.findAll();
     }
 
-    public Producto findById(long id){
+    public Producto findById(Integer id){
         return productoRepository.findById(id).get();
     }
 
@@ -28,7 +28,7 @@ public class ProductoService {
         return productoRepository.save(producto);
     }
 
-    public void delete(Long id){
+    public void delete(Integer id){
         productoRepository.deleteById(id);
     }
